+++ NEWS.md
@@ -3,7 +3,7 @@
 * Introduced support for arbitrary test statistics in `calculate()`. In addition
   to the pre-implemented `calculate(stat)` options, taken as strings, users can
   now supply a function defining any scalar-valued test statistic. See
-  `?calculate()` to learn more.
+  `?calculate()` to learn more (#542).
 
 # infer 1.0.9