+++ NEWS.md
@@ -1,5 +1,8 @@
 # infer (development version)
 
+* Fixed bug where adding `shade_confidence_interval(NULL)` or `shade_p_value(NULL)` 
+  to plots resulted in `list()` rather than the unmodified plot (#566).
+
 * Introduced support for arbitrary test statistics in `calculate()`. In addition
   to the pre-implemented `calculate(stat)` options, taken as strings, users can
   now supply a function defining any scalar-valued test statistic. See